+++ vector/vec2.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Vec2 struct {
-	X float32
-	Y float32
+	X float64
+	Y float64
 }
 
 func (vec *Vec2) Add(otherVec Vec2) Vec2 {
@@ -23,15 +23,15 @@ func (vec *Vec2) Subtract(otherVec Vec2) Vec2 {
 	}
 }
 
-func (vec *Vec2) Multiply(n float32) Vec2 {
+func (vec *Vec2) Multiply(n float64) Vec2 {
 	return Vec2{
 		X: vec.X * n,
 		Y: vec.Y * n,
 	}
 }
 
-func (vec *Vec2) Magnitude() float32 {
-	return float32(math.Sqrt((float64(vec.X) * float64(vec.X)) + (float64(vec.Y) * float64(vec.Y))))
+func (vec *Vec2) Magnitude() float64 {
+	return float64(math.Sqrt((float64(vec.X) * float64(vec.X)) + (float64(vec.Y) * float64(vec.Y))))
 }
 
 func (vec *Vec2) Unit() Vec2 {
